[김유정] ins0320/Bhealthy#6 비로그인시 /post 페이지 접근 금지 설정완료

+++ src/main/java/com/Bhealthy/post/bo/PostBO.java
@@ -80,7 +80,9 @@ public class PostBO {
 		postRepository.delete(post);
 
 		// 이미지 있으면 삭제
-		fileManagerService.deleteFile(post.getImagePath());
+		if( post.getImagePath() != null) {
+			fileManagerService.deleteFile(post.getImagePath());
+		}
 	}
 	
 	// 글 가져오기
